+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres
 to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.12.0] - 2023-06-05
+
+### Changed
+
+* Handling more sql-s the JSqlParser can not parse.
+
+### Added
+
+* A mechanism to disable query parsing through `TwContext` - `TasUtils`.
+
+* Flyway integration for TAS.
+  As JSqlParser is not able to parse DDL queries, we will disable the query parsing for Flyway.
+
 ## [2.11.0] - 2023-05-18
 
 ### Changed